schema sql : sortorder field changed to integer value

+++ schema.sql
@@ -13,5 +13,5 @@ CREATE TABLE IF NOT EXISTS `gantt_tasks` (
   `duration` int(11) NOT NULL,
   `progress` float NOT NULL DEFAULT 0,
   `parent` int(11) NOT NULL,
-  `sortordet` float NOT NULL
+  `sortordet` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
